# swagger-docs/swagger.yaml
openapi: 3.0.0
paths:
...
  /api/driver-manage/driver:
    post:
      operationId: DriverManageController_createDriver
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDriverInput'
      responses:
        '201':
          description: Driver created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseCreateDriverDTO'
      summary: Create a driver
      tags: &ref_0
        - driver-manage
    get:
      operationId: DriverManageController_getAllDriver
      parameters: []
      responses:
        '200':
          description: Get all driver
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseGetAllDriverDTO'
      summary: Get all driver
      tags: *ref_0
...
info:
  title: Shipppee system API
  description: The Shipppee system API description
  version: 1.0.0
  contact: {}
...
servers: []
components:
  schemas:
...
    CreateDriverInput:
      type: object
      properties:
        name:
          type: string
          description: Driver name
          example: Cong Gioi
        email:
          type: string
          description: Driver email
          example: conggioi.pro264@gmail.com
        password:
          type: string
          description: Driver password
          example: '123456'
        phone:
          type: string
          description: Driver phone number
          example: '0123456789'
        transportTypeId:
          type: number
          description: Transport type ID
          example: '1'
      required:
        - name
        - email
        - password
        - phone
        - transportTypeId
    ResponseCreateDriverDTO:
      type: object
      properties:
        code:
          type: number
          example: 200
          description: status code
        success:
          type: boolean
          example: true
          description: status
        message:
          type: string
          example: success
          description: message
        timestamp:
          type: string
          example: '2021-09-01T14:00:00.000Z'
          description: timestamp
        data:
          type: boolean
          example: true
      required:
        - code
        - success
        - message
        - timestamp
        - data
    ResponseGetAllDriverDTO:
      type: object
      properties:
        code:
          type: number
          example: 200
          description: status code
        success:
          type: boolean
          example: true
          description: status
        message:
          type: string
          example: success
          description: message
        timestamp:
          type: string
          example: '2021-09-01T14:00:00.000Z'
          description: timestamp
        data:
          example:
            - name: Cong Gioi
              phone: '0123456789'
              email: abcxyz@gmail.com
              createdAt: '2021-07-07T06:00:00.000Z'
          type: array
          items:
            type: string
      required:
        - code
        - success
        - message
        - timestamp
        - data
...